requirements.txt
Flask==2.0.1
numpy==1.19.5
faiss-cpu==1.7.1.post2
